Extracts prstlib
Moves ble sample to code/nrf-connect/samples
