--- code/nrf-connect/.gitignore
@@ -1,2 +1,0 @@
-build
-build_*
