--- code/nrf-connect/CMakeLists.txt
@@ -1,17 +1,0 @@
-cmake_minimum_required(VERSION 3.20.0)
-
-find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-
-project(b_parasite)
-
-include_directories(src)
-
-target_sources(app PRIVATE
-	src/main.c
-	src/prst/shtc3.c
-	src/prst/adc.c
-	src/prst/led.c
-	src/prst/button.c
-	src/prst/ble/ble.c
-	src/prst/ble/encoding.c
-)
