+++ code/nrf-connect/prstlib/include/prstlib/adc.h
