+++ code/nrf-connect/prstlib/include/prstlib/macros.h
