+++ code/nrf-connect/prstlib/include/prstlib/sensors.h
@@ -1,8 +1,8 @@
 #ifndef _PRST_DATA_H_
 #define _PRST_DATA_H_
 
-#include "prst/adc.h"
-#include "prst/shtc3.h"
+#include "prstlib/adc.h"
+#include "prstlib/shtc3.h"
 
 typedef struct {
   prst_adc_soil_moisture_t soil;
@@ -10,5 +10,7 @@ typedef struct {
   prst_adc_read_t batt;
   prst_shtc3_read_t shtc3;
 } prst_sensors_t;
+
+int prst_sensors_read_all(prst_sensors_t *out);
 
 #endif  // _PRST_DATA_H_
