+++ code/nrf-connect/prstlib/include/prstlib/shtc3.h
