+++ code/nrf-connect/samples/ble/.gitignore
@@ -1,0 +1,3 @@
+build
+build_*
+*.code-workspace
