+++ code/nrf-connect/samples/ble/Kconfig
