+++ code/nrf-connect/samples/ble/boards/arm/bparasite_nrf52840/Kconfig
