+++ code/nrf-connect/samples/ble/boards/arm/bparasite_nrf52840/bparasite_nrf52840_defconfig
