+++ code/nrf-connect/samples/ble/dts/bindings/pwm-fixed.yaml
