+++ code/nrf-connect/samples/ble/nrf52840dk_nrf52840.overlay
