+++ code/nrf-connect/samples/ble/prj.conf
@@ -1,5 +1,5 @@
 # Enabling log has a big impact in power consumption. Only enable it while debugging.
-CONFIG_LOG=n
+CONFIG_LOG=y
 CONFIG_PWM=y
 CONFIG_CBPRINTF_FP_SUPPORT=y
 CONFIG_I2C=y
@@ -24,3 +24,5 @@ CONFIG_CONSOLE=n
 CONFIG_RTT_CONSOLE=n
 
 CONFIG_PINCTRL=y
+
+CONFIG_PRST_SLEEP_DURATION_SEC=600
