+++ code/nrf-connect/samples/ble/src/encoding.h
@@ -1,9 +1,8 @@
 #ifndef _PRST_BLE_ENCODING_H_
 #define _PRST_BLE_ENCODING_H_
 
+#include <prstlib/sensors.h>
 #include <zephyr/bluetooth/bluetooth.h>
-
-#include "prst/data.h"
 
 int prst_ble_encode_service_data(const prst_sensors_t* sensors,
                                  const bt_addr_le_t* bt_addr, uint8_t* out,
